organize the examples into two different projects
