+++ example/rocketmq/src/main/java/org/apache/iotdb/example/Utils.java
